+++ src/application/overwrites/overwriteRequestBody.ts
@@ -1,7 +1,7 @@
 import { PostmanMappedOperation } from '../../postman'
 import { OverwriteRequestBodyConfig } from '../../types'
 import { getByPath, isObject, omitByPath, setByPath } from '../../utils'
-import { FormParam, PropertyList } from 'postman-collection'
+import { FormParam, PropertyList, QueryParam } from 'postman-collection'
 
 /**
  * Overwrite Postman request body with values defined by the portman testsuite
@@ -19,10 +19,17 @@ export const overwriteRequestBody = (
   if (pmOperation.item?.request?.body?.raw) {
     // Overwrite JSON values
     overwriteRequestBodyJson(overwriteValues, pmOperation)
+    return pmOperation
   }
   if (pmOperation.item?.request?.body?.formdata) {
-    // Overwrite Form values
-    overwriteRequestBodyForm(overwriteValues, pmOperation)
+    // Overwrite Form data values
+    overwriteRequestBodyFormData(overwriteValues, pmOperation)
+    return pmOperation
+  }
+  if (pmOperation.item?.request?.body?.urlencoded) {
+    // Overwrite Form url encoded values
+    overwriteRequestBodyFormUrlEncoded(overwriteValues, pmOperation)
+    return pmOperation
   }
 
   return pmOperation
@@ -86,11 +93,11 @@ export const overwriteRequestBodyJson = (
 }
 
 /**
- * Overwrite Formdata values in the request body values
+ * Overwrite form-data values in the request body values
  * @param overwriteValues
  * @param pmOperation
  */
-export const overwriteRequestBodyForm = (
+export const overwriteRequestBodyFormData = (
   overwriteValues: OverwriteRequestBodyConfig[],
   pmOperation: PostmanMappedOperation
   // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types, @typescript-eslint/no-explicit-any
@@ -131,6 +138,10 @@ export const overwriteRequestBodyForm = (
       pmFormParam.disabled = true
     }
 
+    if (overwriteItem?.description !== undefined && pmFormParam?.description) {
+      pmFormParam.description = overwriteItem.description
+    }
+
     // Set Postman form data param
     if (!(overwriteItem?.remove === true)) {
       formMembers.push(pmFormParam)
@@ -162,6 +173,91 @@ export const overwriteRequestBodyForm = (
   // Clear existing & add new members
   formData.clear()
   formData.assimilate(formMembers, false)
+
+  return pmOperation
+}
+
+/**
+ * Overwrite x-www-form-urlencoded values in the request body values
+ * @param overwriteValues
+ * @param pmOperation
+ */
+export const overwriteRequestBodyFormUrlEncoded = (
+  overwriteValues: OverwriteRequestBodyConfig[],
+  pmOperation: PostmanMappedOperation
+  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types, @typescript-eslint/no-explicit-any
+): PostmanMappedOperation => {
+  // Early exit if request body form url encoded is not defined
+  if (!pmOperation.item?.request?.body?.urlencoded) return pmOperation
+  const formEncoded = pmOperation.item.request.body.urlencoded as PropertyList<QueryParam>
+
+  // Early exit if request body form data is empty defined
+  if (formEncoded.count() === 0) return pmOperation
+
+  // New form encoded members
+  const formMembers = [] as QueryParam[]
+
+  const formKeys = formEncoded.map(({ key }) => key)
+  // Detect overwrite form params that do not exist in the Postman collection
+  const insertNewKeys = overwriteValues.filter(x => !formKeys.includes(x.key))
+
+  formEncoded.each(pmFormParam => {
+    // Overwrite values for Keys
+    const overwriteItem = overwriteValues.find(obj => {
+      return obj.key === pmFormParam.key
+    })
+
+    // Test suite - Overwrite/extend request body form encoded param value
+    if (overwriteItem?.value !== undefined && pmFormParam?.value) {
+      const orgValue = pmFormParam.value
+      let newValue = overwriteItem.value
+
+      if (overwriteItem.overwrite === false) {
+        newValue = orgValue + newValue
+      }
+      pmFormParam.value = newValue || 'boolean' === typeof newValue ? `${newValue}`.toString() : ''
+    }
+
+    // Test suite - Disable form encoded param
+    if (overwriteItem?.disable === true) {
+      pmFormParam.disabled = true
+    }
+
+    if (overwriteItem?.description !== undefined && pmFormParam?.description) {
+      pmFormParam.description = overwriteItem.description
+    }
+
+    // Set Postman form encoded param
+    if (!(overwriteItem?.remove === true)) {
+      formMembers.push(pmFormParam)
+    }
+  })
+
+  // Test suite - Add form data param
+  insertNewKeys
+    .filter(overwriteItem => !(overwriteItem.insert === false))
+    .filter(overwriteItem => !(overwriteItem.remove === true))
+    .map(formMember => {
+      // Initialize new Postman query param
+      const pmFormParam = {
+        key: formMember.key,
+        value: '',
+        // description: '',
+        disabled: false
+      } as QueryParam
+
+      // Set form param properties based on the OverwriteValues
+      if (formMember.value) pmFormParam.value = formMember.value
+      if (formMember.disable === true) pmFormParam.disabled = true
+      if (formMember.description) pmFormParam.description = formMember.description
+
+      // Add Postman form encoded param
+      formMembers.push(pmFormParam)
+    })
+
+  // Clear existing & add new members
+  formEncoded.clear()
+  formEncoded.assimilate(formMembers, false)
 
   return pmOperation
 }